21-01-20 22:06

+++ camera.h
@@ -10,10 +10,10 @@ private:
 public:
     __host__ __device__
     camera() {
-        f32 aspect_ratio = 16.0 / 9.0;
-        f32 viewport_height = 2.0;
-        f32 viewport_width = aspect_ratio * viewport_height;
-        f32 focal_length = 1.0;
+        f64 aspect_ratio = 16.0 / 9.0;
+        f64 viewport_height = 2.0;
+        f64 viewport_width = aspect_ratio * viewport_height;
+        f64 focal_length = 1.0;
 
         origin = point3(0, 0, 0);
         horizontal = vec3(viewport_width, 0.0, 0.0);
@@ -21,7 +21,7 @@ public:
         lower_left_corner = origin - horizontal/2 - vertical/2 - vec3(0, 0, focal_length);
     }
     __device__
-    ray get_ray(f32 u, f32 v) const {
+    ray get_ray(f64 u, f64 v) const {
         return ray(origin, lower_left_corner + u*horizontal + v*vertical - origin);
     }
